Capture pieces, multidirectional Capture working in any direction, several directions at once

+++ internal/suggestor/suggestor.go
@@ -3,7 +3,7 @@ package suggestor
 import "gomoku/internal/board"
 import "math/rand"
 
-func SuggestMove() int8 {
-	suggestion := int8(rand.Intn(board.SIZE * board.SIZE))
+func SuggestMove() int {
+	suggestion := int(rand.Intn(board.SIZE * board.SIZE))
 	return suggestion
 }
